+++ autobill/Utils/global_counter.ml
