+++ client/vite.config.ts
@@ -8,5 +8,11 @@ export default defineConfig({
     commonjsOptions: {
       esmExternals: true 
    }
+  },
+  server: {
+    fs: {
+      // Allow serving files from one level up to the project root
+      allow: ['../..'],
+    },
   }
 })
